Fix run issues
