+++ src/admin/client.ts
@@ -10,7 +10,6 @@ var bootstrap = require("../bootstrap.min");
 
 import Models = require("../common/models");
 import moment = require("moment");
-import Exchange = require("./exchange");
 import OrderList = require("./orderlist");
 import Trades = require("./trades");
 import Messaging = require("../common/messaging");
@@ -19,6 +18,9 @@ import Pair = require("./pair");
 import MarketQuoting = require("./market-quoting");
 import MarketTrades = require("./market-trades");
 import Messages = require("./messages");
+import Position = require("./position");
+import Tbp = require("./target-base-position");
+import TradeSafety = require("./trade-safety");
 
 interface MainWindowScope extends ng.IScope {
     env : string;
@@ -115,9 +117,9 @@ var requires = ['ui.bootstrap',
                 MarketQuoting.marketQuotingDirective,
                 MarketTrades.marketTradeDirective,
                 Messages.messagesDirective, 
-                Exchange.positionDirective,
-                Exchange.targetBasePositionDirective,
-                Exchange.tradeSafetyDirective,
+                Position.positionDirective,
+                Tbp.targetBasePositionDirective,
+                TradeSafety.tradeSafetyDirective,
                 Shared.sharedDirectives];
 
 angular.module('projectApp', requires)
