+++ src/service/broker.ts
@@ -12,7 +12,7 @@ import Q = require("q");
 import Interfaces = require("./interfaces");
 import Persister = require("./persister");
 import util = require("util");
-import Messages = require("messages");
+import Messages = require("./messages");
 
 export class MarketDataBroker implements Interfaces.IMarketDataBroker {
     MarketData = new Utils.Evt<Models.Market>();
